Attribute Macro substrate_test_utils::test

#[test]

Marks async function to be executed by an async runtime and provide a TaskExecutor, suitable to test environment.

Requirements

You must have tokio in the [dev-dependencies] of your crate to use this macro.

Example

#[substrate_test_utils::test]
async fn basic_test(task_executor: TaskExecutor) {
    assert!(true);
    // create your node in here and use task_executor
    // then don't forget to gracefully shutdown your node before exit
}